+++ MapController.php
@@ -41,9 +41,9 @@ class MapController extends controller
     function __construct ()
     {
       include_once __DIR__."/models/Game.php";
-      view::set('style_css_path', gila::base_url('src/'.GPACKAGE.'/style.css?v=1012'));
-      view::set('unit_js_path', gila::base_url().'src/'.GPACKAGE.'/unit.js?v=1012');
-      view::set('game_js_path', gila::base_url().'src/'.GPACKAGE.'/gameplay.js?v=1012');
+      view::set('style_css_path', gila::base_url('src/'.GPACKAGE.'/style.css?v=1013'));
+      view::set('unit_js_path', gila::base_url().'src/'.GPACKAGE.'/unit.js?v=1013');
+      view::set('game_js_path', gila::base_url().'src/'.GPACKAGE.'/gameplay.js?v=1013');
 
       $this->gameId = $_COOKIE['gameId'] ?? null;