+++ views/loadData.php
@@ -13,11 +13,11 @@ g.get('<?=$gamedata_url?>/<?=$c->gameId?>/items.json', function(data) {
   updateStats();
   assetLoaded('items');
 })
-g.get('<?=$ppath?>data/itemEnchantments.json?v=1012', function(data) {
+g.get('<?=$ppath?>data/itemEnchantments.json?v=1013', function(data) {
   itemEnchantment = JSON.parse(data);
   assetLoaded('itemEnchantments');
 })
-g.get('<?=$ppath?>data/objects.json?v=1012', function(data) {
+g.get('<?=$ppath?>data/objects.json?v=1013', function(data) {
   objectType = JSON.parse(data);
   assetLoaded('objects');
 })